修复测试时遇到的各种问题

+++ hook.sh
@@ -57,6 +57,12 @@ function git_clone() {
     echo "开始 Clone ${12} $3 分支到 $9$4 文件夹" >> $5 2>&1
 
     git -C $9 clone -b $3 ${12} $4 >> $5 2>&1
+
+    if [ $? != 0 ]; then
+        echo 0
+    else
+        echo 1
+    fi
 }
 
 # 初始化
@@ -83,7 +89,7 @@ if [ $1 == 'pushBranch' ]; then
         echo "分支 $3 处于未激活状态" >> $5 2>&1
 
         # 判断分支空闲
-        isIdle=`grep -oP "^$4 0" t | awk -F " " '{print $1}' | wc -l`
+        isIdle=`grep -oP "^$4 0" $2 | awk -F " " '{print $1}' | wc -l`
         if [ $isIdle -gt 0 ]; then
             echo "切换分支 $3 到激活状态" >> $5 2>&1
 
@@ -96,31 +102,34 @@ if [ $1 == 'pushBranch' ]; then
         else
             # 判断其他分支空闲
             otherIdle=`awk '/^[-0-9A-Za-z]* 0 .*/{print $3" "$1}' $2 | sort -k2 | head -n 1`
-            if [ `echo $otherIdle | wc -l` -gt 0 ]; then
+            hasOtherIdle=`echo $otherIdle | wc -L`
+            if [ $hasOtherIdle -gt 0 ]; then
                 oldDir=`echo $otherIdle | awk -v "PATH=$9" -F " " '{print $2}'`
                 
                 echo "将 $oldDir 重命名为 $9$4" >> $5 2>&1
 
-                mv $oldDir $9$4 >> $5 2>&1
-
-                echo "$4 1" >> $2 2>&1
+                mv $9$oldDir $9$4 >> $5 2>&1 \
+                \
+                && echo "$4 1" >> $2 2>&1
 
                 # 添加 Nginx 配置
 
                 # 更新分支
                 git_pull $@
             else
-                git_clone $@
-
-                echo "$4 1" >> $2 2>&1
+                mkdir -p $9$4 >> $5 2>&1
 
                 # 添加 Nginx 配置
+
+                git_clone $@ \
+                \
+                && echo "$4 1" >> $2 2>&1
             fi
         fi
     fi
 
 # 移除分支
-else if [ $1 == 'delBranch' ]; then
+elif [ $1 == 'delBranch' ]; then
     echo "移除分支：$3 ($4)" >> $5 2>&1
 
     # 设置当前分支为闲置状态
@@ -131,9 +140,12 @@ else if [ $1 == 'delBranch' ]; then
 
     # 移除多余文件夹
     saveCount=`expr ${11} + 1`
-    delBranch=`awk '/^[-0-9A-Za-z]* 0 .*/{print $3" "$1}' $2 | sort -r -k2 | tail -n +$saveCount | awk -F " " '{print $1}'`
-    delDir=`echo "$delBranch" | awk -v "PATH=$9" '{print PATH$1}'`
-    echo "$delDir" | xargs rm -rf
-    echo "$delBranch" | xargs -I {} sed -i '/^{} /d' $2
-    echo "移除多余文件夹：`echo "$delDir" | tr '\n' ' '`" >> $5 2>&1
+    delBranch=`awk '/^[-0-9A-Za-z]* 0 .*/{print $3" "$1}' $2 | sort -r -k2 | tail -n +$saveCount | awk -F " " '{print $2}'`
+
+    if [ `echo $delBranch | wc -L` -gt 0 ]; then
+        delDir=`echo "$delBranch" | awk -v "PATH=$9" '{print PATH$1}'`
+        echo "$delDir" | xargs rm -rf
+        echo "$delBranch" | xargs -I {} sed -i '/^{} /d' $2
+        echo "移除多余文件夹：`echo "$delDir" | tr '\n' ' '`" >> $5 2>&1
+    fi
 fi
